+++ CHANGELOG.md
@@ -46,6 +46,11 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 - Now only accepts a `button` element, or a component that renders a `button` element as a trigger
 - `isInModal` is no longer accepted
 
+#### `Popover` Component
+- The popover no longer closes on user dismissal actions in controlled mode. The
+  `onUserDismiss` callback must now be used to update the state of `isOpen` when the
+  popover is in controlled mode.
+
 #### `Dialog` Component
 - `width` and `height` now do not accept numbers. You must pass the unit type (px, vh, etc). 
 
@@ -61,11 +66,18 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 separator characters between items in an inline-block list. If you'd like to just space
 out items, please use GapList, a new component also added in v9.
 
+<<<<<<< HEAD
 #### Removed components
 - `CountdownButton` (only used by Toast)
 - `DecoratedInput` (use TextInput instead)
 - `IconInput` (use TextInput instead)
 - `Section` (use fds classes instead)
+
+## [8.10] DateInput refs
+
+Added ref props for the `input` and popover content elements:
+- `inputRef`
+- `popoverRef`
 
 ## [8.9] DateInput callbacks